rename exception and add pydocmd to circle2 token service
