+++ RDS/circle2_use_cases/token_service/Pipfile
@@ -7,6 +7,7 @@ verify_ssl = true
 pytest = "*"
 pactman = "*"
 pygount = "*"
+pydoc-markdown = "*"
 
 [packages]
 requests = "*"
