+++ RDS/circle2_use_cases/token_service/requirements-dev.txt
@@ -2,22 +2,31 @@
 attrs==19.3.0
 certifi==2019.11.28
 chardet==3.0.4
+click==7.0
 colorama==0.4.3
 idna==2.8
 importlib-metadata==1.3.0 ; python_version < '3.8'
+jinja2==2.10.3
+livereload==2.6.1
+markdown==3.1.1
+markupsafe==1.1.1
+mkdocs==1.0.4
 more-itertools==8.0.2
 packaging==19.2
 pactman==2.25.0
 pluggy==0.13.1
 py==1.8.0
+pydoc-markdown==2.0.5
 pygments==2.5.2
 pygount==1.0.0
 pyparsing==2.4.5
 pytest==5.3.1
+pyyaml==5.2
 requests==2.22.0
 restnavigator==1.0.1
 semver==2.9.0
 six==1.13.0
+tornado==5.1.1
 unidecode==1.1.1
 uritemplate==3.0.0
 urllib3==1.25.7
