+++ RDS/circle2_use_cases/token_service/src/lib/Exceptions/ServiceException.py
@@ -4,6 +4,10 @@ from lib.Token import Token
 
 
 class ServiceNotFoundError(Exception):
+    """
+    Represents an error, when a service was not found.
+    """
+
     def __init__(self, service: Service, msg=None):
         if msg is None:
             msg = f"{service} not found"
@@ -13,6 +17,10 @@ class ServiceNotFoundError(Exception):
 
 
 class TokenNotFoundError(Exception):
+    """
+    Represents an error, when a token was not found.
+    """
+
     def __init__(self, token: Token, msg=None):
         if msg is None:
             msg = f"{token} not found"
@@ -22,6 +30,10 @@ class TokenNotFoundError(Exception):
 
 
 class UserNotFoundError(Exception):
+    """
+    Represents an error, when an user was not found.
+    """
+
     def __init__(self, user: User, msg=None):
         if msg is None:
             msg = f"{user} not found"
@@ -31,6 +43,10 @@ class UserNotFoundError(Exception):
 
 
 class UserHasTokenAlreadyError(Exception):
+    """
+    Represents an error, when a user has already a token for the service.
+    """
+
     def __init__(self, user: User, token: Token, msg=None):
         if msg is None:
             msg = f"{user} has {token} already."
@@ -41,6 +57,10 @@ class UserHasTokenAlreadyError(Exception):
 
 
 class UserAlreadyRegisteredError(UserNotFoundError):
+    """
+    Represents an error, when a user-id is already registered.
+    """
+
     def __init__(self, user: User, msg=None):
         if msg is None:
             msg = f"{user} already registered."
@@ -48,7 +68,12 @@ class UserAlreadyRegisteredError(UserNotFoundError):
         super(UserAlreadyRegisteredError, self).__init__(msg)
         self.user = user
 
+
 class CodeNotExchangeable(Exception):
+    """
+    Represents an error, when a given code for the oauth workflow could not be exchanged against an access_token.
+    """
+
     def __init__(self, code, service, msg=None):
         if msg is None:
             msg = f"Code {code} coult not be exchanged for {service}"
