+++ RDS/circle3_central_services/token_service/src/lib/Exceptions/ServiceException.py
